requirements.txt
fire
scikit-image
tqdm
pyyaml
numba
progress
opencv-python
einops
matplotlib